Fix hash access in transactions breakdown view
The view was using &:amount syntax which expects method calls on objects,
but the controller returns plain hashes. Changed to use hash access syntax
items.sum { |item| item[:amount] } throughout the view.

This fixes the 500 error when accessing /reports with the message:
'undefined method amount for an instance of Hash'